+++ src/edit.h
@@ -17,6 +17,7 @@ void insert_ch(CodePoint ch);
 void join_lines(void);
 void shift_lines(int count);
 void clear_lines(void);
+void delete_lines(void);
 void new_line(void);
 void format_paragraph(size_t text_width);
 void change_case(char mode);